requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopage==0.5.1
cement==3.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cliff==4.2.0
cmd2==2.4.3
dataclasses-json==0.5.7
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.2.3
Flask-PyMongo==2.3.0
frozenlist==1.3.3
geojson==2.5.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.125
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.2
packaging==23.0
pbr==5.11.1
pluggy==1.0.0
prettytable==3.7.0
pydantic==1.10.7
pydantic-mongo==1.0.1
pymongo==4.3.3
pyowm==3.3.0
pyperclip==1.8.2
PySocks==1.7.1
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
requests-mock==1.10.0
six==1.16.0
SQLAlchemy==1.4.47
stevedore==5.0.0
tenacity==8.2.2
-e git+ssh://git@github.com/acruikshank-wm/testbench.git@49f82da136f60e090f6e9928d36583b2a96271ca#egg=testbench
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
